% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr.sqlite-package.R
\docType{package}
\name{sqlr.sqlite-package}
\alias{sqlr.sqlite}
\alias{sqlr.sqlite-package}
\title{sqlr.sqlite: SQLite Dialect for sqlr}
\description{
Renders the schema representation of the sqlr package as SQLite data definition language, and reflects a live SQLite database back into that representation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nbenn/sqlr}
  \item Report bugs at \url{https://github.com/nbenn/sqlr/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Bennett \email{r@nbenn.ch}

Authors:
\itemize{
  \item Nicolas Bennett \email{r@nbenn.ch}
}

}
\keyword{internal}
